# openapi.yaml
openapi: 3.0.0
info:
  title: Drone Backend API
  description: Smart Surveillance Platform - Drone Video Backend Service
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: User login
      description: Authenticate user with username/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    example: "access-token-sample"
                  refreshToken:
                    type: string
                    example: "refresh-token-sample"
                  expiresIn:
                    type: integer
                    example: 3600
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      roles:
                        type: array
                        items:
                          type: string
        '401':
          description: Invalid credentials
  /api/auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: Obtain new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  expiresIn:
                    type: integer
  /api/cameras:
    post:
      tags:
        - Cameras
      summary: Register new camera
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cameraId
                - siteId
                - label
                - rtspUrl
              properties:
                cameraId:
                  type: string
                  example: "cam-001"
                siteId:
                  type: string
                  example: "site-1"
                label:
                  type: string
                  example: "Front Gate Camera"
                rtspUrl:
                  type: string
                  example: "rtsp://192.168.1.100:554/stream"
                status:
                  type: string
                  enum: [online, offline]
                  example: "online"
      responses:
        '201':
          description: Camera registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Camera'
        '400':
          description: Invalid request

    get:
      tags:
        - Cameras
      summary: List cameras
      security:
        - bearerAuth: []
      parameters:
        - name: siteId
          in: query
          schema:
            type: string
          description: Filter by site ID
      responses:
        '200':
          description: List of cameras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Camera'
  /api/cameras/{id}:
    get:
      tags:
        - Cameras
      summary: Get camera by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Camera details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Camera'
        '404':
          description: Camera not found

    put:
      tags:
        - Cameras
      summary: Update camera
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Camera'
      responses:
        '200':
          description: Camera updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Camera'
        '404':
          description: Camera not found

    delete:
      tags:
        - Cameras
      summary: Delete camera
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Camera deleted
        '404':
          description: Camera not found
...
  /api/onprem/agents:
    post:
      tags:
        - On-Prem Agents
      summary: Register on-prem agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agentId
                - siteId
              properties:
                agentId:
                  type: string
                siteId:
                  type: string
                ip:
                  type: string
                capabilities:
                  type: array
                  items:
                    type: string
                  example: ["transcode", "edge-record", "webrtc-proxy"]
      responses:
        '201':
          description: Agent registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  agentId:
                    type: string
  /api/onprem/agents/{id}/heartbeat:
    post:
      tags:
        - On-Prem Agents
      summary: Agent heartbeat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [healthy, warning, offline]
                cpu:
                  type: number
                  description: CPU usage percentage
                mem:
                  type: number
                  description: Memory usage percentage
                streamsActive:
                  type: integer
                timestamp:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Heartbeat received
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /api/onprem/agents/{id}/capabilities:
    get:
      tags:
        - On-Prem Agents
      summary: Get agent capabilities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent capabilities
          content:
            application/json:
              schema:
                type: object
                properties:
                  agentId:
                    type: string
                  capabilities:
                    type: array
                    items:
                      type: string
  /api/ingest/announce:
    post:
      tags:
        - Ingest
      summary: Announce stream availability from on-prem agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agentId
                - cameraId
              properties:
                agentId:
                  type: string
                cameraId:
                  type: string
                streams:
                  type: array
                  items:
                    type: object
                    properties:
                      protocol:
                        type: string
                        enum: [rtsp, webrtc, hls]
                      url:
                        type: string
                      codec:
                        type: string
                      resolution:
                        type: string
      responses:
        '200':
          description: Stream announced
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean
                  ingestId:
                    type: string
  /api/ingest/signed-url:
    post:
      tags:
        - Ingest
      summary: Generate signed ingest URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agentId
                - cameraId
                - protocol
              properties:
                agentId:
                  type: string
                cameraId:
                  type: string
                protocol:
                  type: string
                  enum: [rtmp, srt, webrtc]
                ttlSeconds:
                  type: integer
                  example: 3600
      responses:
        '200':
          description: Signed URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                  method:
                    type: string
                    enum: [POST, PUT]
                  expiresAt:
                    type: string
                    format: date-time
...
  /api/compare:
    post:
      tags:
        - Comparison
      summary: Create video comparison session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - leftRecordingId
                - rightRecordingId
              properties:
                leftRecordingId:
                  type: string
                rightRecordingId:
                  type: string
                leftStart:
                  type: string
                  format: date-time
                rightStart:
                  type: string
                  format: date-time
      responses:
        '202':
          description: Comparison session queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  compareSessionId:
                    type: string
                  status:
                    type: string
                    example: "queued"
  /api/compare/{id}/status:
    get:
      tags:
        - Comparison
      summary: Get comparison session status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comparison status
          content:
            application/json:
              schema:
                type: object
                properties:
                  compareSessionId:
                    type: string
                  status:
                    type: string
                    enum: [queued, preparing, ready, failed]
                  left:
                    type: object
                    properties:
                      playbackUrl:
                        type: string
                  right:
                    type: object
                    properties:
                      playbackUrl:
                        type: string
  /api/internal/streams/allocate:
    post:
      tags:
        - Internal
      summary: Allocate stream resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cameraId
                - protocol
              properties:
                cameraId:
                  type: string
                agentId:
                  type: string
                targetProfiles:
                  type: array
                  items:
                    type: string
                protocol:
                  type: string
                  enum: [webrtc, hls]
                sessionId:
                  type: string
      responses:
        '201':
          description: Stream allocated
          content:
            application/json:
              schema:
                type: object
                properties:
                  allocationId:
                    type: string
                  proxyEndpoint:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
  /api/internal/streams/release:
    post:
      tags:
        - Internal
      summary: Release stream allocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                allocationId:
                  type: string
      responses:
        '200':
          description: Stream released
  /api/internal/transcode/jobs:
    post:
      tags:
        - Internal
      summary: Create transcode job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source:
                  type: string
                cameraId:
                  type: string
                profiles:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      resolution:
                        type: string
                      bitrate:
                        type: integer
      responses:
        '201':
          description: Transcode job created
  /api/internal/transcode/jobs/{id}/status:
    get:
      tags:
        - Internal
      summary: Get transcode job status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transcode job status
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                  status:
                    type: string
                  progress:
                    type: integer
...
components:
  schemas:
    Camera:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cameraId:
          type: string
        siteId:
          type: string
        label:
          type: string
        rtspUrl:
          type: string
        status:
          type: string
          enum: [online, offline]
        lastSeen:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT