# src/main/java/com/example/droneserver/dto/CameraDto.java
package com.example.droneserver.dto;

public class CameraDto {
    public String id;
    public String cameraId;
    public String siteId;
    public String label;
    public String rtspUrl;
    public String status;
}